requirements.txt
# please make sure you have already a pytorch install that is cuda enabled!
av
einops
flashy>=0.0.1
hydra-core>=1.1
hydra_colorlog
julius
num2words
numpy<2.0.0
sentencepiece
spacy==3.7.6
torch==2.1.0
torchaudio>=2.0.0,<2.1.2
huggingface_hub
tqdm
transformers>=4.31.0  # need Encodec there.
xformers<0.0.23
demucs
librosa
soundfile
gradio
torchmetrics
encodec
protobuf
torchvision==0.16.0
torchtext==0.16.0
# pesq  # Disabled: requires C++ build tools on Windows
pystoi
torchdiffeq
openai
pinecone-client
supabase
elevenlabs
python-dotenv
